_projects/ml-poc/requirements.txt
kumoai
pandas
python-dotenv
